业务代码提交

+++ chair-manager/src/main/java/com/chair/manager/controller/ConsumedDetailsController.java
@@ -27,7 +27,7 @@ public class ConsumedDetailsController {
 	 */
 	@ResponseBody
 	@RequestMapping(value = "/query/{cId}" ,method = RequestMethod.GET)
-	public ConsumedDetails queryOrderById(@PathVariable("cId") Long cId) {
+	public ConsumedDetails queryOrderById(@PathVariable("cId") Integer cId) {
 		return consumedDetailsService.findById(cId);
 	}
